tts-server/requirements.txt
fastapi
uvicorn
qwen-tts
soundfile
numpy
ffmpeg-python
python-multipart
